+++ src/main/java/com/example/yummypet/enums/ItemType.java
@@ -1,5 +1,5 @@
 package com.example.yummypet.enums;
 
 public enum ItemType {
-    product, pet, service
+    product, pet
 }