+++ src/main/java/com/example/yummypet/enums/ServiceStatus.java
@@ -1,6 +1,0 @@
-package com.example.yummypet.enums;
-
-public enum ServiceStatus {
-    pending, in_progress, completed, cancelled
-}
-